pyproject.toml
[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[project]
name = "worktreeflow"
version = "0.3.0"
description = "Git workflow manager for feature branches using worktrees, shows you cli equivalent commands"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Steve Morin", email = "steve.morin@gmail.com" }
]
keywords = ["git", "worktree", "workflow", "cli", "github"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]

dependencies = [
    "gitpython>=3.1.40",
    "click>=8.1.0",
    "rich>=13.0",
    "typing-extensions>=4.0",
]

[project.urls]
Homepage = "https://github.com/smorinlabs/worktreeflow"
Repository = "https://github.com/smorinlabs/worktreeflow"
Issues = "https://github.com/smorinlabs/worktreeflow/issues"

[project.scripts]
wtf = "worktreeflow.cli:cli"
worktreeflow = "worktreeflow.cli:cli"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "ty>=0.0.1a7",
    "pip-audit>=2.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]


[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.coverage.run]
source = ["src/worktreeflow"]

[tool.coverage.report]
fail_under = 35
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]

[tool.hatch.build.targets.wheel]
packages = ["src/worktreeflow"]
